nova_agent/requirements.txt
torch>=2.1.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
datasets>=2.18.0
gymnasium>=0.29.0
minigrid>=2.3.0
wandb>=0.16.0
weave>=0.51.0
wandb-workspaces>=0.1.0
numpy>=1.24.0
mistral-common>=1.4.0
Pillow>=10.0.0
mss>=9.0.0
pyautogui>=0.9.54
pynput>=1.7.6
